lots of changes
Did a lot to the CSS to get the page looking the way I wanted. Made a new model to handel key points on each battle. This required making a api, json view, serializer, url and view.  I added content to <p> tags that still need to be filled in. I set the map page up so that I only had to have one page but that one page would cycle through maps so when the user picks the battle from the drop down the correct map pops up.